Move local stuff back into its own script
