+++ .env.example
@@ -1,5 +1,4 @@
 BRIDGE_CONTROLLER_ADDRESS=0x11241e35B3f79099123aA0C1C4c97b1FcdCd21f6
-DEPLOY_ENVIRONMENT=local
 MNEMONIC="flappy bird couch ... mirror"
 NCT_ADDRESS=0x7beCBA11618Ca63Ead5605DE235f6dD3b25c530E
 POLYGONSCAN_API_KEY=HUC9H2U1S544....
