+++ lib/bridge.js
@@ -1,8 +1,8 @@
 const { ethers } = require("hardhat");
 
-async function deployBridge(regenBridge, nctPool) {
+async function deployBridge(tokenIssuer, nctPool) {
 	const Bridge = await ethers.getContractFactory("ToucanRegenBridge");
-	const bridge = await Bridge.deploy(regenBridge, nctPool);
+	const bridge = await Bridge.deploy(tokenIssuer, nctPool);
 
 	await bridge.deployed();
 	return bridge;
